backend/requirements.txt
# FastAPI and ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication (JWT validation only)
python-jose[cryptography]==3.3.0

# WebSocket
python-socketio==5.11.0
websockets==12.0

# AI Integration
requests==2.31.0
openai>=1.0.0

# Computer Vision
opencv-python==4.9.0.80
mediapipe==0.10.9
numpy==1.26.3

# Audio Processing
SpeechRecognition==3.10.1
librosa==0.10.1
pydub==0.25.1

# File Storage
boto3==1.34.34

# Task Queue
celery==5.3.6
celery[redis]==5.3.6

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
